requirements.txt
flask
jinja2

youtube-transcript-api
yt-dlp
webvtt-py

sentence-transformers
transformers
torch
numpy
tqdm
rank-bm25

google-genai
